delete(show_gui): use new, delete the ole gui framework
