+++ gui/__init__.py
@@ -1,5 +1,4 @@
-from .show_gui import show_gui
 from .refactor_pages.json_file_docker import tab_area
 
 
-__all__ = ["show_gui", "tab_area"]
+__all__ = ["tab_area"]
